Исправлен класс ChangerLibraryService; начато написание классов RemoverLibraryService и FinderLibraryService

+++ app/Services/ChangerLibraryService.php
@@ -2,12 +2,10 @@
 
 namespace App\Services;
 
-use App\Http\Requests\StoreBookRequest;
 use App\Http\Requests\UpdateBookRequest;
 use App\Models\Author;
 use App\Models\AuthorBook;
 use App\Models\Book;
-use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
 
 class ChangerLibraryService
@@ -30,45 +28,25 @@ class ChangerLibraryService
 
         try {
             DB::beginTransaction();
-            
-            try {
-                if (!empty($listOfChangedFieldsExceptAuthors)) {
-                    $this->updateBookInformationExceptAuthors($listOfChangedFieldsExceptAuthors, $formData, $bookId);
-                }
-            } catch (\Exception $e) {
-                echo $e->getMessage();
-                dump($e);
+
+            if (!empty($listOfChangedFieldsExceptAuthors)) {
+                $this->updateBookInformationExceptAuthors($listOfChangedFieldsExceptAuthors, $formData, $bookId);
             }
 
-            try {
-                if ($formData['deletedAuthors'] !== null) {
-                    $this->removeAuthorFromBook($this->helper->getListOfAuthors($formData['deletedAuthors']), $bookId);
-                }
-            } catch (\Exception $e) {
-                echo $e->getMessage();
-                dump($e);
+            $this->changeAuthorsFullName($this->helper->getListOfAuthors($formData['originalOrModifiedAuthors']));
+
+            if ($formData['deletedAuthors'] !== null) {
+                $this->removeAuthorFromBook($this->helper->getListOfAuthors($formData['deletedAuthors']), $bookId);
             }
 
-            try {
-                if ($formData['newAuthors'] !== null) {
-                    $this->addAuthorToBook($this->helper->getListOfAuthors($formData['newAuthors']), $bookId);
-                }
-            } catch (\Exception $e) {
-                echo $e->getMessage();
-                dump($e);
-            }
-
-            try {
-                $this->changeAuthorsFullName($this->helper->getListOfAuthors($formData['originalOrModifiedAuthors']));
-            } catch (\Exception $e) {
-                echo $e->getMessage();
-                dump($e);
+            if ($formData['newAuthors'] !== null) {
+                $this->addAuthorToBook($this->helper->getListOfAuthors($formData['newAuthors']), $bookId);
             }
 
             DB::commit();
         } catch (\Exception $e) {
             DB::rollBack();
-            dump($e->getMessage());
+            dump($e);
         }
     }
 
@@ -104,10 +82,32 @@ class ChangerLibraryService
     private function removeAuthorFromBook(array $authorsToBeDeleted, int $bookId): void
     {
         $idOfAuthorsToBeDeleted = array_map(function ($authorsFullName) {
-            return Author::query()->where('full_name', '=', $authorsFullName)->get('id')->first()->id;
+            $id = Author::query()->where('full_name', '=', $authorsFullName)->get('id')->first();
+
+            if ($id === null) {
+                die('Попытка удалить автора, которого нет в таблице authors');
+            }
+            return $id->id;
         }, $authorsToBeDeleted);
 
+        $idOfAuthorsAssociatedWithBook = array_map(function ($author) {
+            return $author->id;
+        }, Book::find($bookId)->authors()->get(['id'])->all());
+
+        if (count($idOfAuthorsAssociatedWithBook) === 1) {
+            if (!empty(array_intersect($idOfAuthorsToBeDeleted, $idOfAuthorsAssociatedWithBook))) {
+                die('Нельзя удалять автора у книги, если он единственный у нее. Книга не может быть без автора');
+            }
+        }
+
         foreach ($idOfAuthorsToBeDeleted as $authorId) {
+
+            //Сделать проверку, что удаляемый id есть в $idOfAuthorsAssociatedWithBook
+
+            if (!in_array($authorId, $idOfAuthorsAssociatedWithBook)) {
+                die ('Автор отвязывается только от своей книги');
+            }
+
             DB::table('author_book')->where('author_id', '=', $authorId)->where('book_id', '=', $bookId)->delete();
 
             $numberOfBooksWrittenByAuthors = AuthorBook::query()->where('author_id', '=', $authorId)->count();
@@ -125,7 +125,7 @@ class ChangerLibraryService
 
             if (!$numberOfAuthorsWithThisFullName) {
                 $author = new Author();
-                $author->full_name = $authorsFullName;
+                $author->full_name = trim($authorsFullName);
                 $author->save();
             }
 
@@ -143,14 +143,23 @@ class ChangerLibraryService
             if (preg_match('/\([а-яё ]*\)$/ui', $author)) {
 
                 $oldFullName = preg_split('/^[а-яё ]*/ui', $author, -1, PREG_SPLIT_NO_EMPTY);
-                $oldFullName = mb_substr($oldFullName[0], 1, mb_strlen($oldFullName[0])-2);
+                $oldFullName = trim(mb_substr($oldFullName[0], 1, mb_strlen($oldFullName[0])-2));
 
                 $newFullName = preg_split('/\([а-яё ]*\)$/ui', $author, -1, PREG_SPLIT_NO_EMPTY);
+                $newFullName = trim($newFullName[0]);
+
+                //Проверка, что авторов с ФИО === $newFullName нет в таблице authors
+
+                $numberOfAuthorsWithThisFullName = Author::query()->where('full_name', '=', $newFullName)->count();
+
+                if ($numberOfAuthorsWithThisFullName !== 0) {
+                    die('Нельзя изменять ФИО автора на то, которое занято другим автором.');
+                }
 
                 $authorId = Author::query()->where('full_name', '=', $oldFullName)->get('id')->first()->id;
 
                 $authorModel = Author::find($authorId);
-                $authorModel->full_name = $newFullName[0];
+                $authorModel->full_name = $newFullName;
                 $authorModel->save();
             }
         }
